Convert echo to C17, rename zoneinfo header
